chore: move logup evaluator trait to separate file
